backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Hyeonjong <nianelo4@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
djangorestframework-camel-case = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
